Fixed signed/unsigned mismatch warnings

+++ ext/ruby-ffmpeg/ruby_ffmpeg_audio_frame.c
@@ -60,13 +60,13 @@ void audio_frame_mark(void * opaque) {
 VALUE audio_frame_new(AVFrame * frame, AVCodecContext * codec) {
 	// Time stamp: start of with best effort
 	int64_t timestamp = frame->best_effort_timestamp;
-	if (timestamp == AV_NOPTS_VALUE) {
+	if (timestamp == (int64_t)AV_NOPTS_VALUE) {
 		// Fall back to presentation timestamp of frame
 		timestamp = frame->pts;
-		if (timestamp == AV_NOPTS_VALUE) {
+		if (timestamp == (int64_t)AV_NOPTS_VALUE) {
 			// Fall back to presentation timestamp of packet
 			timestamp = frame->pkt_pts;
-			if (timestamp == AV_NOPTS_VALUE) {
+			if (timestamp == (int64_t)AV_NOPTS_VALUE) {
 				// Fall back to decompression timestamp of packet
 				timestamp = frame->pkt_dts;
 			}
@@ -98,8 +98,8 @@ VALUE audio_frame_new(AVFrame * frame, AVCodecContext * codec) {
 						  	codec->sample_fmt,
 						  	codec->frame_size,
 						  	codec->sample_rate,
-						  	(timestamp != AV_NOPTS_VALUE) ? rb_float_new(timestamp * av_q2d(codec->time_base)) : Qnil,
-						  	(duration != AV_NOPTS_VALUE) ? rb_float_new(duration * av_q2d(codec->time_base)) : Qnil);
+						  	(timestamp != (int64_t)AV_NOPTS_VALUE) ? rb_float_new(timestamp * av_q2d(codec->time_base)) : Qnil,
+						  	(duration != (int64_t)AV_NOPTS_VALUE) ? rb_float_new(duration * av_q2d(codec->time_base)) : Qnil);
 }
 
 // Create new instance
